refactor async error hanlder
